subcommandify
